+++ ContractManager/ContractBlueprint/Prerequisite.cs
@@ -31,17 +31,25 @@ namespace ContractManager.ContractBlueprint
         [XmlElement("maxConcurrentCount", DataType = "unsignedInt")]
         public uint maxConcurrentCount { get; set; } = 0;
 
-        // type: hasCompletedContract - offer contract if ...
+        // type: hasCompletedContract - offer contract if the specified contract blueprint uid has been completed.
         [XmlElement("hasCompletedContract", DataType = "string")]
         public string hasCompletedContract { get; set; }
 
-        // type: hasFailedContract - offer contract if ...
+        // type: hasFailedContract - offer contract if the specified contract blueprint uid has been failed.
         [XmlElement("hasFailedContract", DataType = "string")]
         public string hasFailedContract { get; set; }
 
-        // type: hasAcceptedContract - offer contract if ...
+        // type: hasAcceptedContract - offer contract if the specified contract blueprint uid has been accepted (and not yet completed/failed).
         [XmlElement("hasAcceptedContract", DataType = "string")]
         public string hasAcceptedContract { get; set; }
+
+        // type: minNumberOfVessels - offer contract if number of vessels in current celestial system is more than this number.
+        [XmlElement("minNumberOfVessels", DataType = "unsignedInt")]
+        public uint minNumberOfVessels { get; set; } = 0;
+
+        // type: maxNumberOfVessels - offer contract if number of vessels in current celestial system is less than this number.
+        [XmlElement("maxNumberOfVessels", DataType = "unsignedInt")]
+        public uint maxNumberOfVessels { get; set; } = uint.MaxValue;
 
 
         public Prerequisite() { }
@@ -99,9 +107,9 @@ namespace ContractManager.ContractBlueprint
         HasFailedContract,
         [XmlEnum("hasAcceptedContract")]
         HasAcceptedContract,
-        //[XmlEnum("minNumberOfVessels")]
-        //MinNumberOfVessels,
-        //[XmlEnum("maxNumberOfVessels")]
-        //MaxNumberOfVessels
+        [XmlEnum("minNumberOfVessels")]
+        MinNumberOfVessels,
+        [XmlEnum("maxNumberOfVessels")]
+        MaxNumberOfVessels
     }
 }